Allow other primitives as keys
This change makes it possible to use other primitives as keys.
There are a lot of function-level constraints, which could be either
aliased, or lifted to the impl clause, but I prefer to have the
minimal set on a per-function basis.

This is a breaking change, because of the introduction of the extra type
parameter.